Updated README.md
docs: uploaded README.md to include file descriptions and basic instructions

This README contains file descriptions and basic instructions to control the Universal Robot based on detected hand gestures. Instructions on how to operate the Universal Robot is not detailed here as it assumes the user has prior knowledge on how to operate the robot.